Commerce automation improvemenst
share page actions code between plain code approach and builder approach
