+++ ts/packages/agents/browser/src/agent/instacart/actionHandler.mts
@@ -5,29 +5,12 @@ import { ActionContext } from "@typeagent/agent-sdk";
 import { BrowserActionContext } from "../actionHandler.mjs";
 import { BrowserConnector } from "../browserConnector.mjs";
 import { createInstacartPageTranslator } from "./translator.mjs";
-import {
-  AllListsInfo,
-  RecipeInfo,
-  BuyItAgainHeaderSection,
-  BuyItAgainNavigationLink,
-  HomeLink,
-  ListDetailsInfo,
-  ListInfo,
-  ListsNavigationLink,
-  NearbyStoresList,
-  ProductDetailsHeroTile,
-  ProductTile,
-  RecipeHeroSection,
-  SearchInput,
-  StoreInfo,
-  ShoppingCartButton,
-  ShoppingCartDetails,
-  ShoppingCartStoreSection,
-} from "./schema/pageComponents.mjs";
+
 import {
   PurchaseResults,
   PurchaseSummary,
 } from "../commerce/schema/shoppingResults.mjs";
+import { setupPageActions } from "./pageActions.mjs";
 
 export async function handleInstacartAction(
   action: any,
@@ -42,6 +25,8 @@ export async function handleInstacartAction(
     context.sessionContext.agentContext.browserConnector;
 
   const agent = await createInstacartPageTranslator("GPT_4_O_MINI");
+
+  const ui = setupPageActions(browser, agent);
 
   switch (action.actionName) {
     case "searchForProductAction":
@@ -76,109 +61,16 @@ export async function handleInstacartAction(
       break;
   }
 
-  type UIElementSchemas = {
-    AllListsInfo: AllListsInfo;
-    RecipeInfo: RecipeInfo;
-    BuyItAgainHeaderSection: BuyItAgainHeaderSection;
-    BuyItAgainNavigationLink: BuyItAgainNavigationLink;
-    HomeLink: HomeLink;
-    ListDetailsInfo: ListDetailsInfo;
-    ListInfo: ListInfo;
-    ListsNavigationLink: ListsNavigationLink;
-    NearbyStoresList: NearbyStoresList;
-    ProductDetailsHeroTile: ProductDetailsHeroTile;
-    ProductTile: ProductTile;
-    RecipeHeroSection: RecipeHeroSection;
-    SearchInput: SearchInput;
-    StoreInfo: StoreInfo;
-    ShoppingCartButton: ShoppingCartButton;
-    ShoppingCartStoreSection: ShoppingCartStoreSection;
-    ShoppingCartDetails: ShoppingCartDetails;
-  };
-
-  async function getPageComponent<T extends keyof UIElementSchemas>(
-    componentType: T,
-    selectionCondition?: string,
-  ): Promise<UIElementSchemas[T] | undefined> {
-    const htmlFragments = await browser.getHtmlFragments(true);
-    const timerName = `getting ${componentType} section`;
-
-    console.time(timerName);
-    const response = await agent.getPageComponentSchema(
-      componentType,
-      selectionCondition,
-      htmlFragments,
-      undefined,
-    );
-
-    if (!response.success) {
-      console.error("Attempt to get page component failed");
-      console.error(response.message);
-      return undefined;
-    }
-
-    console.timeEnd(timerName);
-    return response.data as UIElementSchemas[T];
-  }
-
-  async function followLink(linkSelector: string | undefined) {
-    if (!linkSelector) return;
-
-    await browser.clickOn(linkSelector);
-    await browser.awaitPageInteraction();
-    await browser.awaitPageLoad();
-  }
-
-  async function searchOnWebsite<T extends keyof UIElementSchemas>(
-    componentType: T,
-    keywords: string,
-  ): Promise<UIElementSchemas[T] | undefined> {
-    if (componentType == "StoreInfo" || componentType == "RecipeInfo") {
-      await goToHomepage();
-    }
-
-    const selector = await getPageComponent("SearchInput");
-    if (!selector) {
-      return;
-    }
-
-    const searchSelector = selector.cssSelector;
-
-    await browser.clickOn(searchSelector);
-
-    let queryPrefix = "";
-    switch (componentType) {
-      case "StoreInfo": {
-        queryPrefix = "stores: ";
-        break;
-      }
-      case "RecipeInfo": {
-        queryPrefix = "recipes: ";
-        break;
-      }
-    }
-
-    await browser.enterTextIn(queryPrefix + keywords, searchSelector);
-    await browser.clickOn(selector.submitButtonCssSelector);
-    await browser.awaitPageInteraction();
-    await browser.awaitPageLoad();
-
-    const request = `Search result: ${keywords}`;
-    const result = await getPageComponent(componentType, request);
-
-    return result as UIElementSchemas[T];
-  }
-
   async function handleFindProduct(action: any) {
-    const targetProduct = await searchOnWebsite(
+    const targetProduct = await ui.searchOnWebsite(
       "ProductTile",
       action.parameters.keyword,
     );
-    await followLink(targetProduct?.detailsLinkSelector);
+    await ui.followLink(targetProduct?.detailsLinkSelector);
   }
 
   async function handleAddToCart(action: any) {
-    const targetProduct = await getPageComponent("ProductDetailsHeroTile");
+    const targetProduct = await ui.getPageComponent("ProductDetailsHeroTile");
 
     if (targetProduct?.addToCartButton) {
       await browser.clickOn(targetProduct.addToCartButton.cssSelector);
@@ -186,12 +78,12 @@ export async function handleInstacartAction(
   }
 
   async function selectStoreCart(action: any) {
-    const cartButton = await getPageComponent("ShoppingCartButton");
+    const cartButton = await ui.getPageComponent("ShoppingCartButton");
     console.log(cartButton);
 
-    await followLink(cartButton?.detailsLinkCssSelector);
+    await ui.followLink(cartButton?.detailsLinkCssSelector);
 
-    const cartDetails = await getPageComponent("ShoppingCartDetails");
+    const cartDetails = await ui.getPageComponent("ShoppingCartDetails");
     console.log(cartDetails);
   }
 
@@ -201,14 +93,14 @@ export async function handleInstacartAction(
   }
 
   async function handleAddToList(action: any) {
-    const targetProduct = await getPageComponent("ProductDetailsHeroTile");
+    const targetProduct = await ui.getPageComponent("ProductDetailsHeroTile");
 
     if (targetProduct?.addToListButton) {
       await browser.clickOn(targetProduct.addToListButton.cssSelector);
 
       // this launches a popup with the available lists
       const request = `ListName: ${action.listName}`;
-      const targetList = await getPageComponent("AllListsInfo", request);
+      const targetList = await ui.getPageComponent("AllListsInfo", request);
 
       if (targetList?.lists) {
         await browser.clickOn(targetList.lists[0].cssSelector);
@@ -218,37 +110,37 @@ export async function handleInstacartAction(
   }
 
   async function goToHomepage() {
-    const link = await getPageComponent("HomeLink");
+    const link = await ui.getPageComponent("HomeLink");
     console.log(link);
 
-    await followLink(link?.linkCssSelector);
+    await ui.followLink(link?.linkCssSelector);
   }
 
   async function handleFindStores(action: any) {
     await goToHomepage();
-    const storesList = await getPageComponent("NearbyStoresList");
+    const storesList = await ui.getPageComponent("NearbyStoresList");
     console.log(storesList);
     return storesList;
   }
 
   async function handleSetPreferredStore(action: any) {
-    const targetStore = await searchOnWebsite(
+    const targetStore = await ui.searchOnWebsite(
       "StoreInfo",
       action.parameters.storeName,
     );
-    await followLink(targetStore?.detailsLinkCssSelector);
+    await ui.followLink(targetStore?.detailsLinkCssSelector);
 
     // TODO: persist preferrences
   }
 
   async function handleFindRecipe(action: any) {
-    const recipe = await searchOnWebsite(
+    const recipe = await ui.searchOnWebsite(
       "RecipeInfo",
       action.parameters.keyword,
     );
 
     if (recipe && recipe.detailsLinkCssSelector) {
-      await followLink(recipe.detailsLinkCssSelector);
+      await ui.followLink(recipe.detailsLinkCssSelector);
     }
   }
 
@@ -260,15 +152,15 @@ export async function handleInstacartAction(
       deliveryInformation: "",
     };
 
-    const recipe = await searchOnWebsite(
+    const recipe = await ui.searchOnWebsite(
       "RecipeInfo",
       action.parameters.recipeName,
     );
 
     if (recipe && recipe.detailsLinkCssSelector) {
-      await followLink(recipe.detailsLinkCssSelector);
+      await ui.followLink(recipe.detailsLinkCssSelector);
 
-      const targetRecipe = await getPageComponent("RecipeHeroSection");
+      const targetRecipe = await ui.getPageComponent("RecipeHeroSection");
 
       if (targetRecipe?.addAllIngridientsCssSelector) {
         await browser.clickOn(targetRecipe.addAllIngridientsCssSelector);
@@ -278,7 +170,6 @@ export async function handleInstacartAction(
         }
 
         const friendlyMessage = await agent.getFriendlyPurchaseSummary(results);
-
         if (friendlyMessage.success) {
           message = (friendlyMessage.data as PurchaseSummary).formattedMessage;
         }
@@ -296,34 +187,27 @@ export async function handleInstacartAction(
 
     await selectStore(action.parameters.storeName);
 
-    const navigationLink = await getPageComponent("ListsNavigationLink");
+    const navigationLink = await ui.getPageComponent("ListsNavigationLink");
     console.log(navigationLink);
 
     if (navigationLink?.linkCssSelector) {
-      await followLink(navigationLink?.linkCssSelector);
+      await ui.followLink(navigationLink?.linkCssSelector);
 
       const request = `List name: ${action.parameters.listName}`;
-      const targetList = await getPageComponent("ListInfo", request);
+      const targetList = await ui.getPageComponent("ListInfo", request);
 
       if (targetList?.detailsLinkCssSelector) {
-        await followLink(targetList.detailsLinkCssSelector);
-
-        const listDetails = await getPageComponent("ListDetailsInfo");
+        await ui.followLink(targetList.detailsLinkCssSelector);
+        const listDetails = await ui.getPageComponent("ListDetailsInfo");
 
         if (listDetails && listDetails.products) {
-          for (let product of listDetails.products) {
-            if (product.addToCartButtonCssSelector) {
-              await browser.clickOn(product.addToCartButtonCssSelector);
-              await browser.awaitPageInteraction();
-              results.addedToCart.push(product);
-            } else {
-              results.unavailable.push(product);
-            }
-          }
+          results = await ui.addAllProductsToCart(
+            listDetails.products,
+            action.parameters.storeName,
+          );
         }
 
         const friendlyMessage = await agent.getFriendlyPurchaseSummary(results);
-
         if (friendlyMessage.success) {
           message = (friendlyMessage.data as PurchaseSummary).formattedMessage;
         }
@@ -334,10 +218,10 @@ export async function handleInstacartAction(
   async function selectStore(storeName: string) {
     await goToHomepage();
     const request = `Store name: ${storeName}`;
-    const targetStore = await getPageComponent("StoreInfo", request);
+    const targetStore = await ui.getPageComponent("StoreInfo", request);
 
     console.log(targetStore);
-    await followLink(targetStore?.detailsLinkCssSelector);
+    await ui.followLink(targetStore?.detailsLinkCssSelector);
   }
 
   async function handleBuyItAgain(action: any) {
@@ -350,34 +234,32 @@ export async function handleInstacartAction(
 
     await selectStore(action.parameters.storeName);
 
-    const navigationLink = await getPageComponent("BuyItAgainNavigationLink");
+    const navigationLink = await ui.getPageComponent(
+      "BuyItAgainNavigationLink",
+    );
 
     console.log(navigationLink);
 
     if (navigationLink) {
-      await followLink(navigationLink.linkCssSelector);
+      await ui.followLink(navigationLink.linkCssSelector);
 
-      const headerSection = await getPageComponent("BuyItAgainHeaderSection");
+      const headerSection = await ui.getPageComponent(
+        "BuyItAgainHeaderSection",
+      );
       console.log(headerSection);
 
       if (headerSection?.products) {
         if (action.parameters.allItems) {
-          for (let product of headerSection.products) {
-            if (product.availability == "Out of stock") {
-              results.unavailable.push(product);
-            } else {
-              if (product.addToCartButtonCssSelector) {
-                await browser.clickOn(product.addToCartButtonCssSelector);
-                await browser.awaitPageInteraction();
-                results.addedToCart.push(product);
-              } else {
-                results.unavailable.push(product);
-              }
-            }
-          }
+          results = await ui.addAllProductsToCart(
+            headerSection?.products,
+            action.parameters.storeName,
+          );
         } else {
           const request = `Product: ${action.productName}`;
-          const targetProduct = await getPageComponent("ProductTile", request);
+          const targetProduct = await ui.getPageComponent(
+            "ProductTile",
+            request,
+          );
           if (targetProduct && targetProduct.addToCartButtonCssSelector) {
             await browser.clickOn(targetProduct.addToCartButtonCssSelector);
             await browser.awaitPageInteraction();
